requirements.txt
openai
pymilvus
python-dotenv
numpy
pymupdf
python-docx
python-pptx